+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+6.1.1
+=====
+
+fix: fix AutoScaling group creation with latest AWS provider
+fix: fix `cpu_options` for ASG
+fix: fix not working dynamic blocks throughout the module
+tech: update possible version for AWS provider to 2.60 and up
+
 6.1.0
 =====